requirements.txt
pyserial==3.5
python-dotenv==1.0.0
slackclient==2.9.4
